Deferred shading, shadows, PBR material improved

+++ VEngineNative/Media.cpp
@@ -16,7 +16,7 @@ char* get_file_contents(const char* path) {
     FILE* fh = fopen(path, "r");
     int size = fsize(fh);
     char* content = (char*)calloc(size + 1, sizeof(char));
-    int sz = fread(content, sizeof(char), size, fh);
+    size_t sz = fread(content, sizeof(char), size, fh);
     content = (char*)realloc(content, sizeof(char) * sz + 1);
     content[sz] = 0;
     fclose(fh);
@@ -27,7 +27,7 @@ int get_file_contents_binary(unsigned char** out_bytes, const char* path) {
     FILE* fh = fopen(path, "rb");
     int size = fsize(fh);
     unsigned char* content = (unsigned char*)calloc(size, 1);
-    int sz = fread(content, 1, size, fh);
+    size_t sz = fread(content, 1, size, fh);
     fclose(fh);
     *out_bytes = content;
     return size;
